More path-name adjstments due to the move
