+++ opencog/attention/HebbianUpdatingAgent.h
@@ -25,11 +25,11 @@
 #define _OPENCOG_HEBBIAN_LEARNING_AGENT_H
 
 #include <string>
+#include <opencog/util/Logger.h>
 
 #include <opencog/atomspace/AtomSpace.h>
 #include <opencog/atomspace/AttentionValue.h>
-#include <opencog/server/Agent.h>
-#include <opencog/util/Logger.h>
+#include <opencog/cogserver/server/Agent.h>
 
 namespace opencog
 {
