+++ opencog/attention/ImportanceSpreadingAgent.h
@@ -30,7 +30,7 @@
 
 #include <opencog/atomspace/AtomSpace.h>
 #include <opencog/atomspace/AttentionValue.h>
-#include <opencog/server/Agent.h>
+#include <opencog/cogserver/server/Agent.h>
 #include <opencog/util/Logger.h>
 
 namespace opencog
