+++ opencog/attention/SimpleImportanceDiffusionAgent.h
@@ -30,7 +30,7 @@
 #include <math.h>
 #include <opencog/atomspace/AtomSpace.h>
 #include <opencog/atomspace/AttentionValue.h>
-#include <opencog/server/Agent.h>
+#include <opencog/cogserver/server/Agent.h>
 #include <opencog/util/Logger.h>
 #include <opencog/util/RandGen.h>
 #include "SpreadDecider.h"
