+++ opencog/cogserver/shell/GenericShell.cc
@@ -22,10 +22,10 @@
 
 #include <thread>
 
-#include <opencog/cogserver/server/ConsoleSocket.h>
 #include <opencog/util/Logger.h>
 #include <opencog/util/platform.h>
 
+#include <opencog/cogserver/server/ConsoleSocket.h>
 #include "GenericEval.h"
 #include "GenericShell.h"
