+++ opencog/cogserver/shell/PythonShellModule.cc
@@ -31,10 +31,11 @@
  */
 #ifdef HAVE_CYTHON
 
-#include <opencog/cython/PythonEval.h>
-#include <opencog/cogserver/server/ConsoleSocket.h>
 #include <opencog/util/Logger.h>
 #include <opencog/util/platform.h>
+
+#include <opencog/cython/PythonEval.h>
+#include <opencog/cogserver/server/ConsoleSocket.h>
 #include "PythonShellModule.h"
 
 namespace opencog
