+++ opencog/cython/opencog/CMakeLists.txt
@@ -134,7 +134,7 @@ IF (HAVE_SERVER)
 	CYTHON_ADD_MODULE_PYX(cogserver
 		"${ATOMSPACE_INCLUDE_DIR}/opencog/cython/opencog/atomspace.pxd"
 		"request.pyx" "agent.pyx" "agent_finder_types.h"
-		"../../server/Agent.h" "../../server/Request.h"
+		"../../cogserver/server/Agent.h" "../../cogserver/server/Request.h"
 		"../../spacetime/SpaceServer.h" "../../spacetime/TimeServer.h"
 	)
 
@@ -163,7 +163,7 @@ IF (HAVE_SERVER)
 	CYTHON_ADD_MODULE_PYX(agent_finder
 		"${ATOMSPACE_INCLUDE_DIR}/opencog/cython/opencog/atomspace.pxd"
 		"cogserver.pxd"
-		"../../server/Agent.h" "../../server/Request.h"
+		"../../cogserver/server/Agent.h" "../../cogserver/server/Request.h"
 		"../../spacetime/SpaceServer.h" "../../spacetime/TimeServer.h"
 	)
