+++ opencog/cython/opencog/cogserver.pxd
@@ -27,14 +27,14 @@ cdef extern from "opencog/atomspace/Handle.h" namespace "opencog":
 
 # AtomSpaces
 from opencog.atomspace cimport cAtomSpace
-cdef extern from "opencog/server/BaseServer.h" namespace "opencog":
+cdef extern from "opencog/cogserver/server/BaseServer.h" namespace "opencog":
     cAtomSpace* server_atomspace()
 
 # ideally we'd import these typedefs instead of defining them here but I don't
 # know how to do that with Cython
 ctypedef short stim_t
 
-cdef extern from "opencog/server/Agent.h" namespace "opencog":
+cdef extern from "opencog/cogserver/server/Agent.h" namespace "opencog":
     cdef cppclass cAgent "opencog::Agent":
         int frequency()
         void setFrequency(int f)
@@ -52,7 +52,7 @@ cdef extern from "opencog/server/Agent.h" namespace "opencog":
 cdef class MindAgent:
     cdef cAgent *c_obj
 
-cdef extern from "opencog/server/Request.h" namespace "opencog":
+cdef extern from "opencog/cogserver/server/Request.h" namespace "opencog":
     cdef cppclass cRequest "opencog::Request":
         void send(string s)
