+++ opencog/nlp/types/NLPTypes.cc
@@ -6,7 +6,7 @@
  * Copyright (c) 2009, 2014 Linas Vepstas <linasvepstas@gmail.com>
  */
 
-#include <opencog/server/Module.h>
+#include <opencog/cogserver/server/Module.h>
 #include "opencog/nlp/types/atom_types.definitions"
 
 #define INHERITANCE_FILE "opencog/nlp/types/atom_types.inheritance"
