+++ opencog/openpsi/OpenPsiModule.cc
@@ -23,7 +23,7 @@
 
 #include "OpenPsiModule.h"
 
-#include <opencog/server/CogServer.h>
+#include <opencog/cogserver/server/CogServer.h>
 
 using namespace opencog;
