+++ opencog/spacetime/atom_types_init.cc
@@ -21,7 +21,7 @@
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 
-#include <opencog/server/Module.h>
+#include <opencog/cogserver/server/Module.h>
 #include "opencog/spacetime/atom_types.definitions"
 
 #define INHERITANCE_FILE "opencog/spacetime/atom_types.inheritance"
